move stray types to typings/ folder
