+++ typings/global.d.ts
